Updated MF2 reference source to last public available version

+++ mf2/src/main/java/minefantasy/mf2/client/render/RenderParachute.java
@@ -12,38 +12,35 @@ import net.minecraft.util.ResourceLocation;
 import org.lwjgl.opengl.GL11;
 
 @SideOnly(Side.CLIENT)
-public class RenderParachute extends Render
-{
+public class RenderParachute extends Render {
     protected ModelBase modelParachute;
 
-    public RenderParachute()
-    {
+    public RenderParachute() {
         this.shadowSize = 1F;
         this.modelParachute = new ModelParachute();
     }
 
     /**
-     * Actually renders the given argument. This is a synthetic bridge method, always casting down its argument and then
-     * handing it off to a worker function which does the actual work. In all probabilty, the class Render is generic
-     * (Render<T extends Entity) and this method has signature public void func_76986_a(T entity, double d, double d1,
-     * double d2, float f, float f1). But JAD is pre 1.5 so doesn't do that.
+     * Actually renders the given argument. This is a synthetic bridge method,
+     * always casting down its argument and then handing it off to a worker function
+     * which does the actual work. In all probabilty, the class Render is generic
+     * (Render<T extends Entity) and this method has signature public void
+     * func_76986_a(T entity, double d, double d1, double d2, float f, float f1).
+     * But JAD is pre 1.5 so doesn't do that.
      */
-    public void doRender(EntityParachute entity, double x, double y, double z, float pitch, float yaw)
-    {
+    public void doRender(EntityParachute entity, double x, double y, double z, float pitch, float yaw) {
         GL11.glPushMatrix();
-        GL11.glTranslatef((float)x, (float)y+1.5F, (float)z);
+        GL11.glTranslatef((float) x, (float) y + 1.5F, (float) z);
         GL11.glRotatef(180.0F - pitch, 0.0F, 1.0F, 0.0F);
-        float f2 = (float)entity.getTimeSinceHit() - yaw;
+        float f2 = entity.getTimeSinceHit() - yaw;
         float f3 = entity.getDamageTaken() - yaw;
 
-        if (f3 < 0.0F)
-        {
+        if (f3 < 0.0F) {
             f3 = 0.0F;
         }
 
-        if (f2 > 0.0F)
-        {
-            GL11.glRotatef(MathHelper.sin(f2) * f2 * f3 / 10.0F * (float)entity.getForwardDirection(), 1.0F, 0.0F, 0.0F);
+        if (f2 > 0.0F) {
+            GL11.glRotatef(MathHelper.sin(f2) * f2 * f3 / 10.0F * entity.getForwardDirection(), 1.0F, 0.0F, 0.0F);
         }
 
         float f4 = 1F;
@@ -57,29 +54,31 @@ public class RenderParachute extends Render
     }
 
     /**
-     * Returns the location of an entity's texture. Doesn't seem to be called unless you call Render.bindEntityTexture.
+     * Returns the location of an entity's texture. Doesn't seem to be called unless
+     * you call Render.bindEntityTexture.
      */
-    protected ResourceLocation getEntityTexture(EntityParachute p_110775_1_)
-    {
+    protected ResourceLocation getEntityTexture(EntityParachute p_110775_1_) {
         return TextureHelperMF.getResource("textures/models/object/parachute.png");
     }
 
     /**
-     * Returns the location of an entity's texture. Doesn't seem to be called unless you call Render.bindEntityTexture.
+     * Returns the location of an entity's texture. Doesn't seem to be called unless
+     * you call Render.bindEntityTexture.
      */
-    protected ResourceLocation getEntityTexture(Entity p_110775_1_)
-    {
-        return this.getEntityTexture((EntityParachute)p_110775_1_);
+    protected ResourceLocation getEntityTexture(Entity p_110775_1_) {
+        return this.getEntityTexture((EntityParachute) p_110775_1_);
     }
 
     /**
-     * Actually renders the given argument. This is a synthetic bridge method, always casting down its argument and then
-     * handing it off to a worker function which does the actual work. In all probabilty, the class Render is generic
-     * (Render<T extends Entity) and this method has signature public void func_76986_a(T entity, double d, double d1,
-     * double d2, float f, float f1). But JAD is pre 1.5 so doesn't do that.
+     * Actually renders the given argument. This is a synthetic bridge method,
+     * always casting down its argument and then handing it off to a worker function
+     * which does the actual work. In all probabilty, the class Render is generic
+     * (Render<T extends Entity) and this method has signature public void
+     * func_76986_a(T entity, double d, double d1, double d2, float f, float f1).
+     * But JAD is pre 1.5 so doesn't do that.
      */
-    public void doRender(Entity p_76986_1_, double p_76986_2_, double p_76986_4_, double p_76986_6_, float p_76986_8_, float p_76986_9_)
-    {
-        this.doRender((EntityParachute)p_76986_1_, p_76986_2_, p_76986_4_, p_76986_6_, p_76986_8_, p_76986_9_);
+    public void doRender(Entity p_76986_1_, double p_76986_2_, double p_76986_4_, double p_76986_6_, float p_76986_8_,
+                         float p_76986_9_) {
+        this.doRender((EntityParachute) p_76986_1_, p_76986_2_, p_76986_4_, p_76986_6_, p_76986_8_, p_76986_9_);
     }
 }
